+++ lib/core/clients/cache/cache_migration_client.dart
@@ -14,17 +14,21 @@ enum MigrationStatus {
 
 class CacheMigrationClient {
   Future<void> migrate({required List<int> encryptionKey}) async {
-    final hasHydratedBox = await _boxExists(CacheConstants.hydratedBox);
-    final migrationStatus = await _getMigrationStatus();
+    try {
+      final hasHydratedBox = await _boxExists(CacheConstants.hydratedBox);
+      final migrationStatus = await _getMigrationStatus();
 
-    if (hasHydratedBox && migrationStatus == MigrationStatus.notCompleted) {
-      await _migrateToEncryptedBox(
-        hydratedBoxName: CacheConstants.hydratedBox,
-        tempBoxName: CacheConstants.tempBox,
-        encryptionKey: encryptionKey,
-      );
-    } else {
-      await _setMigrationStatus(MigrationStatus.completed);
+      if (hasHydratedBox && migrationStatus == MigrationStatus.notCompleted) {
+        await _migrateToEncryptedBox(
+          hydratedBoxName: CacheConstants.hydratedBox,
+          tempBoxName: CacheConstants.tempBox,
+          encryptionKey: encryptionKey,
+        );
+      } else {
+        await _setMigrationStatus(MigrationStatus.completed);
+      }
+    } catch (e) {
+      log('Error during migration: $e');
     }
   }
 
@@ -48,38 +52,36 @@ class CacheMigrationClient {
     required String tempBoxName,
     required List<int> encryptionKey,
   }) async {
-    try {
-      // Open the existing hydrated box.
-      final oldHydratedBox = await Hive.openBox<dynamic>(hydratedBoxName);
-      // Create a new temporary box to hold the data from the hydrated box
-      final tempBox = await Hive.openBox<dynamic>(tempBoxName);
+    // Open the existing hydrated box.
+    final oldHydratedBox = await Hive.openBox<dynamic>(hydratedBoxName);
+    // Create a new temporary box to hold the data from the hydrated box
+    final tempBox = await Hive.openBox<dynamic>(tempBoxName);
 
-      // Migrate data from the existing box to the temporary box
-      for (final key in oldHydratedBox.keys) {
-        final value = oldHydratedBox.get(key);
-        await tempBox.put(key, value);
-      }
-      // Delete the old hydrated box from the disk
-      await oldHydratedBox.deleteFromDisk();
+    // Migrate data from the existing box to the temporary box
+    for (final key in oldHydratedBox.keys) {
+      final value = oldHydratedBox.get(key);
+      await tempBox.put(key, value);
+    }
+    // Delete the old hydrated box from the disk
+    await oldHydratedBox.deleteFromDisk();
 
-      // Open a new hydrated box with the given encryption key
-      final newHydratedBox = await Hive.openBox<dynamic>(
-        hydratedBoxName,
-        encryptionCipher: HiveAesCipher(encryptionKey),
-      );
-      // Migrate the data from the temporary box to the new encrypted box
-      for (final key in tempBox.keys) {
-        final value = tempBox.get(key);
-        await newHydratedBox.put(key, value);
-      }
-      // Complete the migration by closing the new box, deleting the temp box, and updating the migration status
-      await Future.wait([
+    // Open a new hydrated box with the given encryption key
+    final newHydratedBox = await Hive.openBox<dynamic>(
+      hydratedBoxName,
+      encryptionCipher: HiveAesCipher(encryptionKey),
+    );
+    // Migrate the data from the temporary box to the new encrypted box
+    for (final key in tempBox.keys) {
+      final value = tempBox.get(key);
+      await newHydratedBox.put(key, value);
+    }
+    // Complete the migration by closing the new box, deleting the temp box, and updating the migration status
+    await Future.wait(
+      [
         newHydratedBox.close(),
         tempBox.deleteFromDisk(),
         _setMigrationStatus(MigrationStatus.completed),
-      ]);
-    } catch (e) {
-      log('Error during migration: $e');
-    }
+      ],
+    );
   }
 }
